(feat) structure vulpea-create template and enforce property block
Fixes #90

In short, `vulpea` requires each note to have an assigned id.
Documentation [describes][1] a way to automatically assign identifiers
to all notes, but it's not managed by `vulpea` itself, so unless user
configures this piece, or passes template with property block,
`vulpea-create` is overly sensitive.

This change makes `vulpea-create` to accept more structured template,
which is in turn converted into valid org-roam-template and enforces
existence of property block with id value.

[1]: https://github.com/d12frosted/vulpea#mandatory-id-property