(feat) persist meta in database
Fixes #100

The approach is to hook into `org-roam-db` routine. It is not ideal
and intrusive, but provides a quick means to achieve the goal without
writing infrastructure for database management.

`vulpea-note-meta` is an associative list. Every key is a string and
the value is a list of strings. Since semantics of each property are
known only to the end user (e.g. it's a link to some other note and
there can be a single value, etc.) we persist only string. In order to
make it easier to access the data with respect of semantics we need
some API similar to existing meta API, but working on assoc. This will
be covered in the upcoming commit.